src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-root-direction-expected.txt

CONSOLE ERROR: line 8: Uncaught TypeError: root.createShadowRoot is not a function
Element direction details are accessible to shadow root

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
